+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.6.9 (unreleased)
+0.6.9 (2019-10-01)
 ------------------
 
 - Add lag under which compliances were computed.